Milestone-1_project_files/requirements.txt
# ============================================================
# Milestone 1: Research Paper Retrieval Module
# ============================================================

# Core dependencies for paper search and download
semanticscholar>=0.8.0      # Semantic Scholar API client
requests>=2.31.0             # HTTP library for PDF downloads
python-dotenv>=1.0.0        # Environment variable management
tqdm>=4.66.0                # Progress bars for downloads

# Future milestones (already included)
langchain>=0.1.0            # For text processing and LLM chains
langgraph>=0.0.1            # For workflow orchestration
openai>=1.0.0               # OpenAI API client
PyMuPDF4LLM>=0.0.1          # PDF text extraction
gradio>=4.0.0               # UI for the application

# Optional but recommended
certifi>=2023.7.22          # SSL certificate handling
urllib3>=2.0.0              # HTTP client utilities
